moved dinput8 stuff into their own folder. moved asm stuff into their own folder.
